// src/models/models.tsx
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = {
  [K in keyof T]: T[K];
};
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & {
  [SubKey in K]?: Maybe<T[SubKey]>;
};
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & {
  [SubKey in K]: Maybe<T[SubKey]>;
};
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  Time: any;
  Upload: any;
};
// ...
export type AppointmentSearchInput = {
  searchTerm?: InputMaybe<Scalars["String"]>;
  userId?: InputMaybe<Scalars["ID"]>;
  patientId?: InputMaybe<Scalars["ID"]>;
  appointmentStatusId?: InputMaybe<Scalars["String"]>;
  visitTypeId?: InputMaybe<Scalars["String"]>;
  checkInTime?: InputMaybe<Scalars["Time"]>;
};
// ...
export type ConfirmTreatmentOrderInput = {
  treatmentOrderId: Scalars["ID"];
  treatmentId: Scalars["ID"];
  invoiceNo?: InputMaybe<Scalars["String"]>;
  checkInTime: Scalars["Time"];
  roomId: Scalars["ID"];
};
// ...
export type File = {
  __typename?: "File";
  id: Scalars["ID"];
  contentType: Scalars["String"];
  fileName: Scalars["String"];
  extension: Scalars["String"];
  hash: Scalars["String"];
  size: Scalars["Int"];
  createdAt?: Maybe<Scalars["Time"]>;
};
// ...
export type MutationConfirmTreatmentOrderArgs = {
  input: ConfirmTreatmentOrderInput;
};
// ...
export type PaginationInput = {
  page: Scalars["Int"];
  size: Scalars["Int"];
};
// ...
export type Payment = {
  __typename?: "Payment";
  id: Scalars["ID"];
  invoiceNo: Scalars["String"];
  status: Scalars["String"];
  billingId: Scalars["ID"];
  billing: Billing;
  createdAt: Scalars["Time"];
};
// ...
export type QuerySearchAppointmentsArgs = {
  page: PaginationInput;
  input: AppointmentSearchInput;
};
// ...
export type QueryPatientEncounterLimitByUserArgs = {
  userId: Scalars["ID"];
};
// ...
export type Treatment = {
  __typename?: "Treatment";
  id: Scalars["ID"];
  treatmentOrderId: Scalars["ID"];
  patientChartId: Scalars["ID"];
  note?: Maybe<Scalars["String"]>;
  result?: Maybe<Scalars["String"]>;
  rightEyeText?: Maybe<Scalars["String"]>;
  leftEyeText?: Maybe<Scalars["String"]>;
  generalText?: Maybe<Scalars["String"]>;
  treatmentTypeId: Scalars["ID"];
  treatmentType: TreatmentType;
  treatmentTypeTitle: Scalars["String"];
  status: TreatmentStatus;
  receptionNote: Scalars["String"];
  payments: Array<Payment>;
};
// ...
export type TreatmentOrder = {
  __typename?: "TreatmentOrder";
  id: Scalars["ID"];
  patientChartId: Scalars["ID"];
  patientId: Scalars["ID"];
  firstName: Scalars["String"];
  lastName: Scalars["String"];
  phoneNo: Scalars["String"];
  userName: Scalars["String"];
  orderedById?: Maybe<Scalars["ID"]>;
  orderedBy?: Maybe<User>;
  treatments: Array<Treatment>;
  status: TreatmentOrderStatus;
  createdAt?: Maybe<Scalars["Time"]>;
};
// ...
export enum TreatmentOrderStatus {
  Ordered = "ORDERED",
  Completed = "COMPLETED",
}

export enum TreatmentStatus {
  Ordered = "ORDERED",
  Completed = "COMPLETED",
}

export type TreatmentType = {
  __typename?: "TreatmentType";
  id: Scalars["ID"];
  title: Scalars["String"];
  active: Scalars["Boolean"];
  billings: Array<Maybe<Billing>>;
  supplies: Array<Maybe<Supply>>;
};
// ...
export type User = {
  __typename?: "User";
  id: Scalars["ID"];
  firstName: Scalars["String"];
  lastName: Scalars["String"];
  userTypes: Array<Maybe<UserType>>;
  active: Scalars["Boolean"];
  email: Scalars["String"];
  confirmed?: Maybe<Scalars["Boolean"]>;
  locked?: Maybe<Scalars["Time"]>;
  signature?: Maybe<File>;
  profilePic?: Maybe<File>;
  createdAt?: Maybe<Scalars["Time"]>;
};
// ...
export type UserType = {
  __typename?: "UserType";
  id: Scalars["ID"];
  title: Scalars["String"];
};